+++ crates/icy_board_engine/src/icy_board/state/mod.rs
@@ -52,6 +52,7 @@ pub mod ppl_terminal;
 pub mod ppl_terminal_control;
 pub mod ppl_terminal_info;
 pub mod ppl_terminal_input;
+pub mod ppl_user;
 pub mod user_commands;
 pub mod virtual_screen;
 use self::functions::display_flags;